add sdk to handle gas calculations
