+++ libs/coin-modules/coin-vechain/package.json
@@ -104,7 +104,8 @@
     "invariant": "^2.2.2",
     "lodash": "^4.17.21",
     "rxjs": "^7.8.1",
-    "@vechain/sdk-core": "1.2.0"
+    "@vechain/sdk-core": "2.0.0",
+    "@vechain/sdk-network": "2.0.0"
   },
   "devDependencies": {
     "@faker-js/faker": "^9.6.0",
