+++ libs/coin-modules/coin-vechain/src/common-logic/fixPrefixCase.ts
@@ -1,0 +1,5 @@
+const PREFIX_REGEX = /^0[xX]/;
+
+export const fixPrefixCase = (hex: string): string => {
+  return PREFIX_REGEX.test(hex) ? hex.replace(PREFIX_REGEX, "0x") : `${"0x"}${hex}`;
+};
