+++ libs/coin-modules/coin-vechain/src/common-logic/index.ts
@@ -1,6 +1,8 @@
-export * from "./address-utils";
-export * from "./hex-utils";
-export * from "./logic";
-export * from "./mapping-utils";
-export * from "./pad-address";
-export * from "./transaction-utils";
+export * from "./fixPrefixCase";
+export * from "./calculateClauses";
+export * from "./mapTransfersToOperations";
+export * from "./padAddress";
+export * from "./calculateTransactionInfo";
+export * from "./calculateGasFees";
+export * from "./generateNonce";
+export * from "./parseAddress";
