+++ libs/coin-modules/coin-vechain/src/common-logic/mapTransfersToOperations.ts
