+++ libs/coin-modules/coin-vechain/src/common-logic/padAddress.ts
