+++ libs/coin-modules/coin-vechain/src/constants/env.ts
@@ -1,0 +1,3 @@
+import { getEnv } from "@ledgerhq/live-env";
+
+export const VECHAIN_NODE_URL = getEnv("API_VECHAIN_THOREST");
