+++ libs/coin-modules/coin-vechain/src/constants/index.ts
@@ -1,0 +1,1 @@
+export * from "./env";
