+++ libs/coin-modules/coin-vechain/src/contracts/constants.ts
@@ -2,10 +2,3 @@ export const TransferEventSignature =
   "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef";
 
 export const VTHO_ADDRESS = "0x0000000000000000000000000000456e65726779";
-
-// params.sol - builtin contract (https://www.vechain.org/whitepaper/#bit_v48i3)
-export const PARAMS_ADDRESS = "0x0000000000000000000000000000506172616d73";
-export const REWARD_RATIO_KEY =
-  "0x00000000000000000000000000000000000000007265776172642d726174696f";
-export const BASE_GAS_PRICE_KEY =
-  "0x000000000000000000000000000000000000626173652d6761732d7072696365";
