+++ libs/coin-modules/coin-vechain/src/types/bridge.ts
@@ -43,4 +43,6 @@ export type TransactionInfo = {
   tokenAccount: TokenAccount | undefined;
   estimatedFees: string;
   estimatedGas: number;
+  maxFeePerGas: number;
+  maxPriorityFeePerGas: number;
 };
