+++ libs/coin-modules/coin-vechain/src/types/constants.ts
@@ -1,5 +1,2 @@
 export const MAINNET_CHAIN_TAG = 74;
 export const TESTNET_CHAIN_TAG = 39;
-export const DEFAULT_GAS_COEFFICIENT = 0;
-export const HEX_PREFIX = "0x";
-export const ZERO_ADDRESS = "0x0000000000000000000000000000000000000000";
